--- pytest.ini
@@ -1,6 +1,0 @@
-[pytest]
-# TODO: add coverage report
-addopts = --pep8
-testpaths = tests
-pep8ignore =
-    tests/*.py ALL
