+++ {{cookiecutter.project_slug}}/Dockerfile
