+++ {{cookiecutter.project_slug}}/bootstrap.sh
