+++ {{cookiecutter.project_slug}}/fabfile.py
