+++ {{cookiecutter.project_slug}}/pytest.ini
@@ -1,0 +1,1 @@
+[pytest]
