+++ {{cookiecutter.project_slug}}/requirements/base.txt
@@ -1,0 +1,1 @@
+tornado==4.3
