+++ {{cookiecutter.project_slug}}/requirements/local.txt
@@ -1,0 +1,5 @@
+# Local development dependencies go here
+-r base.txt
+
+ipdb
+Sphinx
