+++ {{cookiecutter.project_slug}}/requirements/production.txt
@@ -1,0 +1,3 @@
+# Production specific requirements go here (they should not vary from the base
+# requirements)
+-r base.txt
