+++ {{cookiecutter.project_slug}}/setup.cfg
