+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/handlers/base.py
