+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/templates/base.html
