+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/urls.py
