+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/version.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+__version__ = '{{ cookiecutter.version }}'
